# docs/swagger.yaml
basePath: /
definitions:
  category.Category:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      role_id:
        type: integer
      updated_at:
        type: string
    type: object
  category.RequestCategory:
    properties:
      name:
        type: string
      role_id:
        type: integer
    type: object
  category.RequestUpdateCategory:
    properties:
      id:
        type: integer
      name:
        type: string
      role_id:
        type: integer
    type: object
...
  category.ResponseAnny:
    properties:
      code:
        type: integer
      data: {}
      error:
        type: boolean
      msg:
        type: string
      type:
        type: integer
    type: object
  category.ResponseCategory:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/category.Category'
      error:
        type: boolean
      msg:
        type: string
      type:
        type: integer
    type: object
  ingredient.Ingredient:
    properties:
      created_at:
        type: string
      id:
        type: integer
      id_product:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  ingredient.RequestIngredient:
    properties:
      id_product:
        type: integer
      name:
        type: string
    type: object
  ingredient.RequestUpdateIngredient:
    properties:
      id:
        type: integer
      id_product:
        type: integer
      name:
        type: string
    type: object
...
  ingredient.ResponseAnny:
    properties:
      code:
        type: integer
      data: {}
      error:
        type: boolean
      msg:
        type: string
      type:
        type: integer
    type: object
  ingredient.ResponseIngredient:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/ingredient.Ingredient'
      error:
        type: boolean
      msg:
        type: string
      type:
        type: integer
    type: object
...
host: 172.147.77.149:5050
info:
  contact:
    email: luis.lucero@bjungle.net
    name: API Support
  description: Api para la gestión de la comida y recetas
  license:
    name: Software Owner
    url: https://www.bjungle.net/terms/licenses
  termsOfService: https://www.bjungle.net/terms/
  title: API FOOD
  version: "1.4"
paths:
  /api/v1/user/category:
    post:
      consumes:
      - application/json
      description: Método que permite crear una categoría de comida
      parameters:
      - description: Datos de la categoría
        in: body
        name: RequestCategory
        required: true
        schema:
          $ref: '#/definitions/category.RequestCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/category.ResponseCategory'
      summary: Método que permite crear una categoría de comida
      tags:
      - Category
    put:
      consumes:
      - application/json
      description: Método que permite actualizar una categoría de comida
      parameters:
      - description: Datos de la categoría
        in: body
        name: RequestUpdateCategory
        required: true
        schema:
          $ref: '#/definitions/category.RequestUpdateCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/category.ResponseCategory'
      summary: Método que permite actualizar una categoría de comida
      tags:
      - Category
  /api/v1/user/category/{id}:
    delete:
      consumes:
      - application/json
      description: Método que permite borrar una categoría de comida
      parameters:
      - description: Id de la categoría
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/category.ResponseAnny'
      summary: Método que permite borrar una categoría de comida
      tags:
      - Category
    get:
      consumes:
      - application/json
      description: Método que permite obtener una categoría de comida por su id
      parameters:
      - description: Id de la categoría
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/category.ResponseCategory'
      summary: Método que permite obtener una categoría de comida por su id
      tags:
      - Category
...
  /api/v1/user/ingredient:
    post:
      consumes:
      - application/json
      description: Método que permite crear un ingrediente de comida
      parameters:
      - description: Datos del ingrediente
        in: body
        name: RequestIngredient
        required: true
        schema:
          $ref: '#/definitions/ingredient.RequestIngredient'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ingredient.ResponseIngredient'
      summary: Método que permite crear un ingrediente de comida
      tags:
      - Ingredient
    put:
      consumes:
      - application/json
      description: Método que permite actualizar un ingrediente de comida
      parameters:
      - description: Datos del ingrediente
        in: body
        name: RequestUpdateIngredient
        required: true
        schema:
          $ref: '#/definitions/ingredient.RequestUpdateIngredient'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ingredient.ResponseIngredient'
      summary: Método que permite actualizar un ingrediente de comida
      tags:
      - Ingredient
  /api/v1/user/ingredient/{id}:
    delete:
      consumes:
      - application/json
      description: Método que permite borrar un ingrediente de comida
      parameters:
      - description: Id del ingrediente
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ingredient.ResponseAnny'
      summary: Método que permite borrar un ingrediente de comida
      tags:
      - Ingredient
    get:
      consumes:
      - application/json
      description: Método que permite obtener un ingrediente de comida por su id
      parameters:
      - description: Id del ingrediente
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ingredient.ResponseIngredient'
      summary: Método que permite obtener un ingrediente de comida por su id
      tags:
      - Ingredient
...
swagger: "2.0"
...